requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.6.0
dj-database-url==0.5.0
Django==3.2.7
django-ckeditor==6.4.2
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-nested-admin==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
filters==1.3.2
fonttools==4.33.3
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
pydyf==0.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pyphen==0.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-monkey-business==1.0.0
pytils==0.4.1
pytz==2022.1
regex==2022.6.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.1.0
zopfli==0.2.1
